AI_Auditor/requirements.txt
pydantic>=2.0
networkx>=3.0
radon>=6.0
rich>=13.0
ollama>=0.1.0
pyyaml>=6.0
tenacity>=8.0
